Don't attempt to fault-in HLSL symbols after a fatal error is found. (#366)
Sema::InstantiatingTemplate::InstantiatingTemplate is backing off at
this point, and so the HLSL built-ins that have the expectation of
succeeding are not going to deal well with this case. In the (common)
case of a file missing, it might be interesting to continue, but for
robustness the first fix to be made is to be resilient under these
conditions and simply back out.