Fix Windows SDK check in dxexp.cpp (#704)
dxexp.cpp tries to define some types if they are not found in
the SDK. However, we cannot use #ifndef to test whether a type
is defined or not. Use WDK_NTDDI_VERSION instead to do the
check instead.

d3d12.h in Windows SDK 10.0.16299.0 defined the types we want in
dxexp.cpp, but the one in Windows SDK 10.0.15063.0 does not.
WDK_NTDDI_VERSION is set to NTDDI_WIN10_RS3 for the former, and
NTDDI_WIN10_RS2 for the later.