[linux-port] Fix conflict with /U arg and files (#1440)
Due to the format that Windows arguments take, files on Unix systems
can sometimes be mistaken for them. A notable example of this is a
file from a directory starting with /U because it will look to the
arg processor exactly like the undefine-macro argument. Because of
this, the VerifierTest might seem to hang because, seeming to have
no input file, it falls back to reading stdin.

The dxc binary avoids this by using a completely different set of
options that doesn't include any of these. clang itself uses
inclusion and exclusion flags that determine on which systems and
under what circumstances / arguments are allowed.

CompilerInvocation::CreateFromArgs was altered to not take any of
these bitfields for the sake of the HLSL tests. I've made a slight
alteration to that to exclude the cl.exe style arguments using the
exclusion bitfield.