Implement Length property on constant arrays (FXC regression) (#1517)
Implement Length property on constant arrays (FXC regression)

The FXC compiler supports Length property on arrays. This change adds it
to the DXC compiler for backwards compatibility. However, it is marked
obsolete and in HLSL Version 2016 will report a warning. On higher versions
it will result in an error to prompt the developer to either change the code 
or to use the -HV flag for backwards compatibility.

Fix HLSLVersion not initialized by default

Adds RewriterTest::RunArrayLength tests which is currently disabled until we have 
a way to set HLSLVersion on RewriterTests.