Return error on subscript access of index vector (#2837)
When an output index array vector is accessed by element instead of the
whole, DXIL generation must produce an error. Previously, when the
access was performed using the dot operator, the error was produced, but
the processing wasn't terminated, so a second error erroneously saying
that the vector had no member by that name was also produced by falling
through to the struct processing. When the index was performed using a
numerical subscript, no error was produced at all and a crash resulted
when the indices were processed later.

By changing the interface for functions that process records that are
really HLSL Matrix, Vector or Array types, inapplicable structs can be
distinguished from expressions that result in an error and the error can
be returned, preventing the duplicate error.

A check for use of the output indices is performed in the subscript
processing to produce an error in this case.

As part of this, an incidental change was made to short-circuit
processing when an array is found to have a member access that isn't
Length.

Tests are added for all the above.

A very incidental change was made to GetStructuralForm() that was
applied to a duplicate function, but never propagated correctly.