Preserve precise metadata when replacing instructions in transformations (#3010)
We were not correctly preserving precise metadata when replacing instructions
in various transformations. This caused precise metadata to be dropped
when one a precise instruction was replaced with a non-precise instruction.

For example, given the two equivalent mul expressions

    float x = mul(a, b)
    precise float y = mul(a, b)
    z = x + y

GVN would replace `y` with `x` in the computation of `z`. This caused x to
be computed in non-precise mode.

The fix is to make sure we combine known dxil metadata when replacing instructions.