Don't reprocess instr in CullSensitiveBlocks (#3029)
Previously, only PHI instructions were collected and prevented from
repeating. This was the minimum required to prevent infinite compiles,
but in complicated shaders with a lot of wave dependencies that feed in
through multiple operands, compilation can take a very long time.

By including all encountered instructions, no repetition of traversal is
performed. The cost is lookups into a much larger set, but that seems
the vastly more performant option.

By collecting call instructions and grouping them by innermost
containing loop, the collection of visited instructions can be shared
since the result of the test for any break loop containing the wave loop
will have the same result so long as the wave loop and instruction
belonging to the break loop are the same.

The instigating shader doesn't actually make use of this optimization,
so I added a test that does.